feat: send 404 report to customers
