+++ src/support/notfound.js
@@ -10,7 +10,23 @@
  * governing permissions and limitations under the License.
  */
 import RUMAPIClient from '@adobe/spacecat-shared-rum-api-client';
-import { build404InitialSlackMessage, build404SlackMessage } from './slack.js';
+import { isArray } from '@adobe/spacecat-shared-utils';
+import commaNumber from 'comma-number';
+import { markdown, section } from './slack.js';
+
+export const INITIAL_404_SLACK_MESSAGE = '*404 REPORT* for the *last week* :thread:';
+
+export function build404InitialSlackMessage(mentions) {
+  return [
+    {
+      type: 'section',
+      text: {
+        type: 'mrkdwn',
+        text: `${isArray(mentions) ? `${mentions.join(' ').toString()} ` : ''}${INITIAL_404_SLACK_MESSAGE}`,
+      },
+    },
+  ];
+}
 
 export const sendInitial404Message = async (slackClient, slackContext) => {
   const blocks = build404InitialSlackMessage(slackContext?.mentions);
@@ -22,6 +38,7 @@ export const sendInitial404Message = async (slackClient, slackContext) => {
   );
   return { thread_ts: threadId, channel: slackContext.channel };
 };
+
 export const get404Backlink = async (context, url) => {
   try {
     const rumApiClient = RUMAPIClient.createFrom(context);
@@ -32,6 +49,37 @@ export const get404Backlink = async (context, url) => {
   }
 };
 
+export function build404SlackMessage(url, auditResult, backlink, mentions) {
+  const blocks = [];
+
+  blocks.push(section({
+    text: markdown(`${isArray(mentions) ? mentions.join(' ').toString() : ''} For *${url}*, ${auditResult.length} page(s) had 404s *last week* for the real users.\n More information is below (up to three pages):`),
+  }));
+
+  for (let i = 0; i < Math.min(3, auditResult.length); i += 1) {
+    const topLine = section({
+      text: markdown(`:arrow-red2: *<${auditResult[i].url}|${auditResult[i].url}>*`),
+    });
+
+    const stats = section({
+      fields: [
+        markdown(`:mag: *Pageviews:* ${commaNumber(auditResult[i].pageviews)}`),
+        markdown(`:mag: *Source:* ${commaNumber(auditResult[i].source)}`),
+      ],
+    });
+
+    blocks.push(topLine);
+    blocks.push(stats);
+  }
+
+  if (backlink) {
+    blocks.push(section({
+      text: markdown(`*To access the full report <${backlink}|click here> :link:* _(expires in 7 days)_`),
+    }));
+  }
+
+  return blocks;
+}
 export const send404Report = async ({
   context,
   slackClient,
